+++ src/main/java/com/jar100/mssproduct/controller/dto/ProductUpdateRequest.java
@@ -5,7 +5,6 @@ import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotNull;
 
 public record ProductUpdateRequest(
-    @NotNull Long id,
     String name,
     Long brandId,
     Category category,
