+++ src/main/java/com/jar100/mssproduct/domain/summary/repository/PriceQueryRepository.java
@@ -1,5 +1,6 @@
 package com.jar100.mssproduct.domain.summary.repository;
 
+import com.jar100.mssproduct.common.dto.ProductChangedEvent;
 import com.jar100.mssproduct.domain.summary.dto.BrandTotalPrice;
 import com.jar100.mssproduct.domain.summary.dto.CategoryMinPrice;
 import com.jar100.mssproduct.domain.summary.dto.CategoryPriceRange;
@@ -15,4 +16,5 @@ public interface PriceQueryRepository {
 
     // 3) 카테고리별 가격 범위
     CategoryPriceRange findPriceRangeByCategory(String categoryName);
+
 }
