+++ src/main/resources/application.yml
@@ -1,6 +1,6 @@
 spring:
   datasource:
-    url: jdbc:h2:mem:product;DB_CLOSE_DELAY=-1;TRACE_LEVEL_SYSTEM_OUT=2
+    url: jdbc:h2:mem:product;DB_CLOSE_DELAY=-1;
     driverClassName: org.h2.Driver
     username: sa
     password:
